data_engineering/HandyFacts/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.2
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==20.1.0
idna==3.4
numpy==1.26.0
pandas==2.1.1
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
